requirements.txt
InstagramApi
requests
parse
emoji
python-dateutil
